Rename css files to use critical path css inlining
