+++ src/app/views/home/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 
-import styles from './styles.scss';
+import styles from './styles.critical.scss';
 
 export default function Home() {
     return (
